backend/requirements.txt
flask
transformers
nltk
torch
python-dotenv